requirements.txt
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
cryptography==35.0.0
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap4==4.0.2
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
paramiko==2.8.0
pycparser==2.20
PyNaCl==1.4.0
PyYAML==6.0
six==1.16.0
SQLAlchemy==1.4.26
visitor==0.1.3
Werkzeug==2.0.2
WTForms==2.3.3
